+++ mingle/utilities/chisqr.py
@@ -2,6 +2,7 @@
 
 import numpy as np
 from joblib import Parallel, delayed
+from scipy.stats import chi2
 
 
 def chi_squared(observed, expected, error=None, axis=0):
@@ -67,3 +68,13 @@ def parallel_chisqr(iter1, iter2, observation, model_func, model_params, n_jobs=
                                                                  model_func, (a, b, *model_params))
                                    for a in iter1 for b in iter2)
     return np.asarray(grid)
+
+
+def chi2_at_sigma(sigma, df=1):
+    """Use inverse survival function to calculate the chi2 value for significances.
+
+    Updated values from https://en.wikipedia.org/wiki/Normal_distribution#Cumulative_distribution_function
+    """
+    sigma_percent = {0: 0, 1: 0.682689492137, 2: 0.954499736104, 3: 0.997300203937,
+                     4: 0.999936657516, 5: 0.999999426697, 6: 0.999999998027}
+    return chi2(df).isf(1 - sigma_percent[sigma])
